+++ include/event_dispatcher/event_dispatcher.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "event_dispatcher/batch_publish_result.hpp"
 #include "event_dispatcher/config.hpp"
 #include "event_dispatcher/dispatcher.hpp"
 #include "event_dispatcher/lifecycle.hpp"